format the output appropiately for python use
Depending on if we want binary or CSV, the output will have different
type of headers.

For instance for binary, there is no header but we need to write a
pgcopy header that allow the binary insertion of the data.

For CSV we want a header, just to make debugging easier if there is an
issue.

Finally move the Python extension code to the top of the file to
mimimize code footprint.